jwt authentication and clundinary image upload for books added

+++ project/api/views/book_views.py
@@ -1,4 +1,7 @@
 from rest_framework import viewsets, status
+from rest_framework.permissions import IsAuthenticated
+from rest_framework.parsers import MultiPartParser, FormParser
+from rest_framework_simplejwt.authentication import JWTAuthentication
 from rest_framework.response import Response
 from rest_framework.decorators import action
 from rest_framework.exceptions import NotFound
@@ -9,6 +12,23 @@ from django.shortcuts import get_object_or_404
 
 class BookViewSet(viewsets.ViewSet):
     queryset = Book.objects.all()
+    parser_classes = [MultiPartParser, FormParser]
+    authentication_classes = [JWTAuthentication]
+    # permission_classes = [IsAuthenticated]  # Require login for all actions
+    print(authentication_classes)
+
+    def check_admin_permission(self, request):
+        """Helper function to check if the user is an admin."""
+        if not request.user.is_staff:
+            return Response(
+                {
+                    "success": False,
+                    "status": 403,
+                    "message": "Only admin users can perform this action.",
+                },
+                status=status.HTTP_403_FORBIDDEN,
+            )
+        return None  # Return None if the user is an admin
 
     def list(self, request):
         books = Book.objects.all()
@@ -24,6 +44,11 @@ class BookViewSet(viewsets.ViewSet):
         )
 
     def create(self, request):
+        # Only admin can create books
+        permission_response = self.check_admin_permission(request)
+        if permission_response:
+            return permission_response
+
         serializer = BookSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -59,6 +84,10 @@ class BookViewSet(viewsets.ViewSet):
         )
 
     def update(self, request, pk=None):
+        permission_response = self.check_admin_permission(request)
+        if permission_response:
+            return permission_response
+        # PUT request: Only admins can update books
         book = get_object_or_404(Book, bookId=pk)
         serializer = BookSerializer(book, data=request.data)
         if serializer.is_valid():
@@ -82,10 +111,14 @@ class BookViewSet(viewsets.ViewSet):
         )
 
     def destroy(self, request, pk=None):
+        # DELETE request: Only admins can delete books
+        permission_response = self.check_admin_permission(request)
+        if permission_response:
+            return permission_response
+
         book = get_object_or_404(Book, bookId=pk)
         book.delete()
         return Response(
             {"success": True, "status": 200, "message": "Book successfully deleted"},
-            status=status.HTTP_200_OK
+            status=status.HTTP_200_OK,
         )
-
